+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for the companion plugin
+- Added support for putting extension settings directly under `lsp.luau-lsp.settings`.
+- Added support for text objects ([docs](https://zed.dev/docs/vim#treesitter)).
 
 ## [0.2.2] - 2024-12-22